+++ lib/libesp32/berry/src/be_strlib.h
@@ -16,6 +16,7 @@ extern "C" {
 #endif
 
 bstring* be_strcat(bvm *vm, bstring *s1, bstring *s2);
+bstring* be_strmul(bvm *vm, bstring *s1, bint n);
 int be_strcmp(bstring *s1, bstring *s2);
 bstring* be_num2str(bvm *vm, bvalue *v);
 void be_val2str(bvm *vm, int index);